Move `Addr` to `TxData`
